+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+## 0.23.1 - Unreleased
+
 ## 0.23.0 - 2026-06-09
 
 ### Added